Add a test driver that works with mboxes.
This is similar in spirit to timtest, but it works with any old kind
of mailbox recognized by the Python mailbox module.

One non-trivial difference from timtest: Rather than requiring that
the user split the mailbox into separate parts, it selects NSETS
different subsets of the mailbox to use for testing.  It chooses an
arbitrary subset because my mailboxes are sorted by date, and I didn't
want to bias tests by choosing training data from a small period of
time.

The timtest module has grown a Driver() class that is intended to work
just like the drive() function, but with a bit more flexibility.  The
jdrive() function might be able to replace drive(), but I can't test
it so I'm not going to replace it.  Maybe Tim will try jdrive() and
report if it works correctly.

I didn't find the MsgStream() class useful outside of timtest, but
mailboxes are represented by the mbox class, which is an iterable
collection of Msg objects.

Renamed the path attribute of Msg to tag, since path doesn't make
sense with an mbox.  The path was getting used as a human-readable tag
for messages, so I synthesized one for mbox messages.