Fixed a bug in the opening of a folder given with "+foo" (wasn't using _factory).
Add a -u option similar to that of GBayes.py.  For this, factored the
opening of the mbox out of train() into a separate function getmbox(),
and the formatting of the clues out of filter().

(The -u option needs work; it currently doesn't report the message
number in a very useful way.)