Add two new options: extract_dow and generate_time_buckets.  I see an ever-so-slight improvement using a 10x10 cv grid w/ 200 ham and spam in each set.  Checking in so others can test it.
The extract_dow option causes tokens like dow:0 (which is Monday) to be
generated.  The generate_time_buckets option causes tokens like time:12:3 to
be generated.  (12 is the hour and 3 indicates the minutes was in the fourth
six-minute bucket.)  I've done nothing to test the bucket granularity.
Lawyers bill in six-minute increments.  Seemed good enough for me. ;-)