Major change in how options work.  No longer do we use the ugly getattr hack, and no longer do we ignore section names. We also have access to lots more information about options, including some information about valid values.
At the moment this includes code to make it 100% backwards
compatible - this will go sooner or later.  I will gradually commit
all the modules that need changing; for the moment nothing
changes.

Note that some options are renamed (all those that included the
section name in the option name have had it removed).