Fix a comment typo in timtest.py
Add urlslurper.py.  This is a version of timtest.py that will retrieve
text from (http) urls included in messages, if the message doesn't
have enough clues to classify it as ham/spam.  This is a standalone
module and does not require modification of any other modules.

To test, run urlsluper.py and compare the output to output from timtest.py.
The docstring has more information.