Update readme.txt to include imapfilter information, and to reflect the recent Options.py changes.
Correct a comment in Outlook2000\export.py

Hopefully fix Options.py so that the temp file problem on
linux goes away (someone tell me if this is not the case)

Improve Options.py so that options use the same delimiter
as when read on export, where possible.

Renamed/moved several options, as discussed on the list.
For the moment, you can still refer to these by the old names
(no modules have been updated to use the new names yet).
The convert_config_file script should also convert these (but
this is as yet untested).  This should be the last major change of
Options.py necessary for some time (I hope!).

Patched hammiefilter so that it still uses the same default
filename for the storage database.  THIS WILL FAIL IF THE USER
HAS CHOSEN TO NAME THE FILE "hammie.db" AND OPTIONS.PY
DOES THE IMPORT OF THEIR CONFIG FILE.  To fix, undo the
patch, or simply put a full pathname in the config file.