Change "Anti-Spam" on the dropdown to "SpamBayes"
Huge changes to configuration.  No longer use a pickle, but instead a series
of .INI files (in the format of the default bayes file - but *not* using
the bayes ini file).  We automatically migrate the old pickle file to the
new INI file.

This is very cool as it allows us to support options before we have a UI
for it.  A first such option is a "data_directory", which would allow
site administrators to specify a different per-user directory.  Some
other options are likely to appear soon.

Source-code users will note the creation of "unknown_profile.ini" - this is
because released win32alls currently can not determine the Outlook profile
name. When it can, this file will magically be renamed to the correct profile
name. The next binary release will know the profile name immediately, so will
never see "unknown_profile.ini".  (This is used so SpamBayes can
eventually be used with multiple profiles - currently it fails pretty
badly)