This is a working draft of the 'overkill' (thanks Greg) app that works like pop3proxy, but allows training via drag and drop in many MUAs (including OE and Mozilla Mail, but not Eudora) courtesy of a local IMAP4 server.  (Note that the name will hopefully change at some point).
I'm checking this as a standalone script at the moment, so you won't be able to just run it.  To do so, you'll need to:
 * install the latest cvs of twisted.
 * add the extra options to Options.py (or ask me for a copy of my one)
 * get a copy of twisted/protocols/imap4.py from me that doesn't have all the bugs of the cvs version, or wait for the cvs version to be fixed.

At the moment, it seems to mostly work with Mozilla mail, although I haven't done any thorough testing yet.  It works ok with Eudora, except that Eudora doesn't let you view two accounts side-by-side, so it's useless.  For some reason messages won't show up in OE at the moment - if anyone has a clue why, I'd love to know.

For whoever it was that was interested in writing an IMAP server for spambayes,
here is code that you may be able to borrow.