If we can't find a config file anywhere but the windows app data directory, then load it.
Swap various over to correct places in prep for 1.0a6.  If you were using notate_to, notate_subject, hammie:debug_header, or many of the pop3proxy storage options, you will need to update your config file.  These are in the most suitable places now.

Use urllib not urllib2 to shut down the proxy.  Why?  Because it works and the other doesn't.

Create the server strings for the ui *after* reading in the command line parameters.

Fix importing Options.py in ProxyUI.py for the binary.