Loosen constraints on HEADER_VALUE regular expression.
Add an experimental/deprecation capability.  (This is more difficult to
describe than to use...)  Options which are experimental or deprecated are
prefixed with 'X-'.  The distinction between the two rather artificial.  The
description of deprecated options should start with "(DEPRECATED)".

There are four cases:

    * option is foo, user sets foo -  Set foo silently (current behavior)

    * option is X-foo, user sets X-foo - Set X-foo silently (current
      behavior)

    * option is foo, user sets X-foo - option has moved from experimental to
      accepted, but the user still sets the option using the experimental
      form.  Set foo silently.

    * option is X-foo, user sets foo - option has moved from accepted to
      deprecated, but the user still sets the non-prefixed option.  Set
      X-foo and emit a warning.

Of course, this should all be considered experimental for the time being and
is subject to change... <wink>