Bring pspam into the modern spambayes world.  update.py, pop.py and scoremsg.py all seem to work ok for me, with very limited testing.
Let the connection address be either a INET address (in the form "(server,port)", or a UNIX one.

Standardise the options code (move the options into the same place as all the rest, and dump
the pspam only file, which didn't work with the new OptionsClass anyway).  Note that the
options have no nice name or docstring, because I don't really know that I know enough to
fill those out.  They might vanish sometime anyway.