Add a docstring.
Add -f command line arg to pass a filename rather than reading from stdin
(which is still the default)

Add a training_is_ham line to the error graph which shows the percentage
of training data that is ham (i.e. shows the imbalance).

Modify the outputing so that it can be in different formats for those of us without
plotmtv.  The -c command line option outputs all the lines in the same set of rows,
rather than in their own set as is the default.  The -s arg specifies the separator
for this sort of output (defaults to a comma, so that csv files are output).

Comparing output from this in default mode and the previous mkgraph.py output
gives me identical results except for one extra newline at the top, so this should
still work with plotmtv, I think (but can't test more than that).