Well, it's been some time, and no-one disagreed, and three people liked the idea, so here it is.
MAY BREAK THINGS!

From now, path/file options are not relative to the current working directory,
they are relative to the last configuration file loaded.  This still defaults to the
current working directory (except on Windows with win32all, where it's still
the User Application Data directory).  So this should mean no noticeable change,
unless your last configuration file is not in the current working directory, and
your data is.