Fix [ 940155 ] Data Folder location
The "show data folder" button in the Manager dialog brought up the Windows
data directory, rather than the SpamBayes one (usually the same, but not always).

It's arguable which one is correct, but I think the most common use for this button
is to find the databases and configuration files, and so data_directory rather than
windows_data_directory is what we are after.  All that would be in the other one
(that is used) is the default_configuration.ini file (if it's not in the SpamBayes program
directory), and anyone using that can find it by hand, I hope.

If you disagree which folder, Mark, at least this is easy to back out ;)