Wow, this is very out of date!
Python 2.3 is no longer in beta...

Use "SpamBayes" not "spambayes".

Update the "what needs to be done" bit.  Please everyone feel free to
edit this to reflext what you think needs to be done.  (Or maybe there
should just be links to the wiki or spambayes-dev or the feature
requests?)