In the questions about setting options there are some python commands in the form 'python -c "command; command"'.  This is a nice form in that the user can just run them from a shell and not need to start up the interactive interpreter, but make for very long lines.  Because they were so long, the whole FAQ would need horizontal scrolling on low resolutions.
That seemed a high price to pay for the more correct way to display the
options, so I've wrapped them, and put in notes explaining that the wrapping
is only for display.  I imagine these particular two FAQs are rarely used now
anyway.

If anyone feels that this is the wrong decision, feel free to back the changes
out :)